+++ src/main/java/step2/service/LottoRank.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.service;
 
 public enum LottoRank {
